chore: move dashboards API towards feature folders
